requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
fastapi==0.115.6
h11==0.14.0
httptools==0.6.4
idna==3.10
oauthlib==3.2.2
osrparse==6.0.2
ossapi==5.0.4
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
sniffio==1.3.1
starlette==0.41.3
typing-utils==0.1.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1

httpx~=0.28.1